move model package inside book package
