+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/model/PredefinedShelfName.java
@@ -1,4 +1,4 @@
-package com.karankumar.bookproject.model;
+package com.karankumar.bookproject.book.model;
 
 import com.karankumar.bookproject.ExcludeFromJacocoGeneratedReport;
