+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/model/README.md
