+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/model/Tag.java
@@ -15,7 +15,7 @@
    If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.karankumar.bookproject.model;
+package com.karankumar.bookproject.book.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.HashSet;
