+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/repository/BookRepository.java
@@ -17,7 +17,7 @@
 
 package com.karankumar.bookproject.book.repository;
 
-import com.karankumar.bookproject.model.Book;
+import com.karankumar.bookproject.book.model.Book;
 import com.karankumar.bookproject.shelf.model.PredefinedShelf.ShelfName;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
