+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/repository/TagRepository.java
@@ -23,7 +23,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.karankumar.bookproject.model.Tag;
+import com.karankumar.bookproject.book.model.Tag;
 
 import java.util.List;
 import java.util.Optional;
