+++ backend/book-app/src/main/java/com/karankumar/bookproject/book/service/BookService.java
@@ -22,12 +22,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.karankumar.bookproject.book.dto.BookPatchDto;
-import com.karankumar.bookproject.model.Author;
-import com.karankumar.bookproject.model.Book;
-import com.karankumar.bookproject.model.BookFormat;
-import com.karankumar.bookproject.model.BookGenre;
+import com.karankumar.bookproject.book.model.Author;
+import com.karankumar.bookproject.book.model.Book;
+import com.karankumar.bookproject.book.model.BookFormat;
+import com.karankumar.bookproject.book.model.BookGenre;
 import com.karankumar.bookproject.shelf.model.PredefinedShelf.ShelfName;
-import com.karankumar.bookproject.model.Publisher;
+import com.karankumar.bookproject.book.model.Publisher;
 import com.karankumar.bookproject.shelf.model.Shelf;
 import com.karankumar.bookproject.book.repository.BookRepository;
 import com.karankumar.bookproject.shelf.service.PredefinedShelfService;
