+++ backend/book-app/src/main/java/com/karankumar/bookproject/shelf/service/UserCreatedShelfService.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.shelf.service;
 
 import com.karankumar.bookproject.account.service.UserService;
-import com.karankumar.bookproject.model.Book;
+import com.karankumar.bookproject.book.model.Book;
 import com.karankumar.bookproject.shelf.repository.UserCreatedShelfRepository;
 import java.util.HashSet;
 import java.util.List;
