+++ backend/book-app/src/main/java/com/karankumar/bookproject/statistics/GenreStatistics.java
@@ -17,9 +17,9 @@
 
 package com.karankumar.bookproject.statistics;
 
-import com.karankumar.bookproject.model.Book;
-import com.karankumar.bookproject.model.BookGenre;
-import com.karankumar.bookproject.model.RatingScale;
+import com.karankumar.bookproject.book.model.Book;
+import com.karankumar.bookproject.book.model.BookGenre;
+import com.karankumar.bookproject.book.model.RatingScale;
 import com.karankumar.bookproject.shelf.service.PredefinedShelfService;
 
 import java.util.ArrayList;
