+++ backend/book-app/src/main/java/com/karankumar/bookproject/statistics/YearStatistics.java
@@ -17,8 +17,8 @@
 
 package com.karankumar.bookproject.statistics;
 
-import com.karankumar.bookproject.model.Book;
-import com.karankumar.bookproject.model.RatingScale;
+import com.karankumar.bookproject.book.model.Book;
+import com.karankumar.bookproject.book.model.RatingScale;
 import com.karankumar.bookproject.shelf.service.PredefinedShelfService;
 
 import java.time.LocalDate;
