+++ backend/book-app/src/test/java/com/karankumar/bookproject/book/service/BookServiceTest.java
@@ -17,8 +17,8 @@
 
 package com.karankumar.bookproject.book.service;
 
-import com.karankumar.bookproject.model.Author;
-import com.karankumar.bookproject.model.Book;
+import com.karankumar.bookproject.book.model.Author;
+import com.karankumar.bookproject.book.model.Book;
 import com.karankumar.bookproject.shelf.model.PredefinedShelf;
 import com.karankumar.bookproject.account.model.User;
 import com.karankumar.bookproject.book.repository.BookRepository;
