+++ backend/book-app/src/test/java/com/karankumar/bookproject/book/service/UserCreatedShelfServiceTest.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.book.service;
 
 import com.karankumar.bookproject.account.service.UserService;
-import com.karankumar.bookproject.model.PredefinedShelfName;
+import com.karankumar.bookproject.book.model.PredefinedShelfName;
 import com.karankumar.bookproject.shelf.model.UserCreatedShelf;
 import com.karankumar.bookproject.account.model.User;
 import com.karankumar.bookproject.shelf.repository.UserCreatedShelfRepository;
