+++ backend/book-app/src/test/java/com/karankumar/bookproject/bookimport/ImportServiceTest.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.bookimport;
 
 import com.karankumar.bookproject.annotations.IntegrationTest;
-import com.karankumar.bookproject.model.Book;
+import com.karankumar.bookproject.book.model.Book;
 import com.karankumar.bookproject.book.service.BookService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
