+++ backend/book-app/src/test/java/com/karankumar/bookproject/statistics/GenreStatisticsTest.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.statistics;
 
 import com.karankumar.bookproject.annotations.IntegrationTest;
-import com.karankumar.bookproject.model.BookGenre;
+import com.karankumar.bookproject.book.model.BookGenre;
 import com.karankumar.bookproject.book.service.BookService;
 import com.karankumar.bookproject.shelf.service.PredefinedShelfService;
 import com.karankumar.bookproject.statistics.util.StatisticTestUtils;
