+++ backend/book-app/src/test/java/com/karankumar/bookproject/statistics/PageStatisticsTest.java
@@ -18,8 +18,8 @@
 package com.karankumar.bookproject.statistics;
 
 import com.karankumar.bookproject.annotations.IntegrationTest;
-import com.karankumar.bookproject.model.Author;
-import com.karankumar.bookproject.model.Book;
+import com.karankumar.bookproject.book.model.Author;
+import com.karankumar.bookproject.book.model.Book;
 import com.karankumar.bookproject.shelf.model.PredefinedShelf;
 import com.karankumar.bookproject.book.service.BookService;
 import com.karankumar.bookproject.shelf.service.PredefinedShelfService;
